requirements.txt
langchain
langchain-openai
langchain-qdrant
langchain-ollama
qdrant-client
fastembed
fastapi==0.115.8
uvicorn
requests
python-dotenv
psutil>=6.0.0
zenml[local]==0.94.1
schedule
mcp